% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{calModuleScore}
\alias{calModuleScore}
\title{The function to calculate module score given gene modules}
\usage{
calModuleScore(object, gene.use = "hubgene")
}
\arguments{
\item{object}{a spMOCA object}

\item{gene.use}{the features used to calculate module scores. The default is hub genes.}
}
\value{
Return a location-by-module module score matrix.
}
\description{
The function to calculate module score given gene modules
}
