% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA.R
\name{createSpatialKernel}
\alias{createSpatialKernel}
\title{The function to create a spatial dependency matrix given spatial coordiante}
\usage{
createSpatialKernel(
  spatialCoord,
  normCount,
  kernelType = "Gaussian",
  kernelBw = "Rule-Of-Thumb"
)
}
\arguments{
\item{spatialCoord}{a location-by-2 spatial coordinate matrix}

\item{normCount}{a gene-by-location normalized matrix count}

\item{kernelType}{the type of Spatial Kernel to be constructed. The default is Gaussian}

\item{kernelBw}{for constructing a spatial kernel, which type of bandwidth estimation will be used. The default is Rule-Of-Thumb}
}
\value{
Return a spatial kernel matrix
}
\description{
The function to create a spatial dependency matrix given spatial coordiante
}
