% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{createSPMOCAObject}
\alias{createSPMOCAObject}
\title{The function to create a spMOCA object for inferring gene co-expression network for SRT}
\usage{
createSPMOCAObject(spatial_countMat, spatial_location, feature_list)
}
\arguments{
\item{spatial_countMat}{a SRT gene-by-locatiion raw count matrix}

\item{spatial_location}{a location-by-2 spatial coordinate data fra,e}

\item{feature_list}{a list of selected features}
}
\value{
Return a spMOCA object with a raw count data, a location data and initial selected features
}
\description{
The function to create a spMOCA object for inferring gene co-expression network for SRT
}
