% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{estimatespMOCA}
\alias{estimatespMOCA}
\title{The function to estimate the gene correlation by maximizing log-likelihood given a spatial dependency}
\usage{
estimatespMOCA(object)
}
\arguments{
\item{object}{a spMOCA object}
}
\value{
Return a spMOCA object with maximized likelihood estimated gene-gene correlation.
}
\description{
The function to estimate the gene correlation by maximizing log-likelihood given a spatial dependency
}
