% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{featureSelectionSPMOCA}
\alias{featureSelectionSPMOCA}
\title{The function to perform a feature selection}
\usage{
featureSelectionSPMOCA(
  object,
  feature.selection = "SPARK",
  feature.list = NULL
)
}
\arguments{
\item{object}{a spMOCA object}

\item{feature.selection}{Feature selection method. The default is SPARK for spatial variable genes selection. 
If feature.selection == Null, no selection will be performed.}

\item{feature.list}{A list of pre-specified selected features. 
Users can pre-specify their features list and updated the spMOCA object directly.}
}
\value{
Return a spMOCA object with updated selected features
}
\description{
The function to perform a feature selection
}
