% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{geneSetAnalysis}
\alias{geneSetAnalysis}
\title{The function to perform gene set analysis}
\usage{
geneSetAnalysis(object, gene.use = "hubgene", fgsea_sets)
}
\arguments{
\item{object}{a spMOCA object}

\item{gene.use}{the features used to calculate module scores. The default is hub genes.}

\item{fgsea_sets}{a list of gene set}
}
\value{
Return a list of odds ratio and p-value tables for each module per each gene set
}
\description{
The function to perform gene set analysis
}
