% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matmult}
\alias{matmult}
\title{A matrix multiplication function written by RCpp to accelerate computation.}
\usage{
matmult(A, B)
}
\arguments{
\item{matrixA}{A matrix}

\item{matrixB}{Another matrix}
}
\value{
the matrix multiplication of A and B
}
\description{
A matrix multiplication function written by RCpp to accelerate computation.
}
