% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{moduleDetectionspMOCA}
\alias{moduleDetectionspMOCA}
\title{The function to perform gene module detection}
\usage{
moduleDetectionspMOCA(object, module.method = "WGCNA", res_parameter = 400)
}
\arguments{
\item{object}{a spMOCA object}

\item{module.method}{module detection method. The default is WGCNA}

\item{res_parameter}{resolution parameter for module detection method}
}
\value{
Return a maximized likelihood gene coexpression correlation estimation.
}
\description{
The function to perform gene module detection
}
