% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{normalizeSPMOCA}
\alias{normalizeSPMOCA}
\title{The function to normalize the count matrix in a spMOCA object}
\usage{
normalizeSPMOCA(object, norm.method = "logNorm")
}
\arguments{
\item{object}{a spMOCA object}

\item{norm.method}{Normalization method. The default is a log-transformation with library size normalization}
}
\value{
Return a spMOCA object with a normalized expression matrix
}
\description{
The function to normalize the count matrix in a spMOCA object
}
