% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMGM.R
\name{spMGMest}
\alias{spMGMest}
\title{The function to estimate the gene covariance by maximizing log-likelihood given a spatial covariance}
\usage{
spMGMest(normCount, spatialKernel)
}
\arguments{
\item{normCount}{a gene-by-spot normalized count matrix}

\item{spatialKernel}{a spatial kernel}
}
\value{
Return a maximized likelihood gene correlation estimation.
}
\description{
The function to estimate the gene covariance by maximizing log-likelihood given a spatial covariance
}
