\name{spMOCA-package}
\alias{spMOCA-package}
\alias{spMOCA}
\docType{package}
\title{
\packageTitle{spMOCA}
}
\description{
\packageDescription{spMOCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spMOCA}
\packageIndices{spMOCA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spMOCA}

Maintainer: \packageMaintainer{spMOCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
