% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA.R
\name{spMOCAest}
\alias{spMOCAest}
\title{The function to estimate the gene correlation by maximizing log-likelihood given a spatial dependency}
\usage{
spMOCAest(normCount, spatialKernel)
}
\arguments{
\item{normCount}{a gene-by-location normalized count matrix}

\item{spatialKernel}{a spatial kernel}
}
\value{
Return a maximized likelihood gene coexpression correlation estimation.
}
\description{
The function to estimate the gene correlation by maximizing log-likelihood given a spatial dependency
}
