% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMGM.R
\name{createSpatialKernel}
\alias{createSpatialKernel}
\title{The function to estimate the gene covariance by maximizing log-likelihood given a spatial covariance}
\usage{
createSpatialKernel(
  spatialCoord,
  normCount,
  kernelType = "Gaussian",
  kernelBw = "Rule-Of-Thumb"
)
}
\arguments{
\item{spatialCoord}{a spot-by-2 spatial coordinate matrix}

\item{normCount}{a gene-by-spot normalized matrix count}

\item{kernelType}{the type of Spatial Kernel to be constructed. The default is Gaussian}

\item{kernelBw}{for constructing a spatial kernel, which type of bandwidth estimation will be used. The default is Rule-Of-Thumb}
}
\value{
Return a spatial kernel matrix
}
\description{
The function to estimate the gene covariance by maximizing log-likelihood given a spatial covariance
}
