% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMOCA_MainFunction.R
\name{createSpatialKernel}
\alias{createSpatialKernel}
\title{The function to create a spatial dependency matrix given spatial coordiante}
\usage{
createSpatialKernel(
  object,
  kernelType = "Gaussian",
  kernelBw = "Rule-Of-Thumb"
)
}
\arguments{
\item{object}{a spMOCA object}

\item{kernelType}{the type of Spatial Kernel to be constructed. The default is Gaussian}

\item{kernelBw}{for constructing a spatial kernel, which type of bandwidth estimation will be used. The default is Rule-Of-Thumb}
}
\value{
Return a spMOCA object a spatial kernel matrix
}
\description{
The function to create a spatial dependency matrix given spatial coordiante
}
